Bug 1080803 - Tighten up AES decryption in ClearKey CDM - r=cpearce

+++ media/gmp-clearkey/0.1/ClearKeyUtils.h
@@ -9,7 +9,7 @@
 #include <string>
 #include <vector>
 
-#define CLEARKEY_KEY_LEN 16
+#define CLEARKEY_KEY_LEN ((size_t)16)
 
 #if 0
 void CK_Log(const char* aFmt, ...);
